pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lightlygpt"
version = "1.0.0"
description = "AI-powered image analysis tool built in collaboration with Lightly.ai"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]

dependencies = [
    "streamlit>=1.25.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    # "clip-by-openai>=1.0",  
    "transformers>=4.30.0",
    "Pillow>=9.5.0",
    "numpy>=1.23.5",
    "pandas>=1.5.3",
    "scikit_learn>=1.2.2",
    "matplotlib>=3.7.1",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "chromadb>=0.4.18",
    "openai==0.28.0",
    "langchain>=0.0.267",
    "langchain-community>=0.0.10",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
    "geopy==2.4.1",
    "opencv-python>=4.5.0",
    "ImageHash>=4.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[project.scripts]
lightly-gpt = "lightlygpt.run:main"
